Change tests to CloudWatch rename also
