+++ events/codepipeline_test.go
@@ -1,99 +1,38 @@
+// Copyright 2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 package events
 
 import (
 	"encoding/json"
-	"github.com/stretchr/testify/require"
 	"io/ioutil"
 	"testing"
-	"time"
+
+	"github.com/aws/aws-lambda-go/events/test"
+	"github.com/stretchr/testify/assert"
 )
 
-func TestUnmarshalCodePipelineEvent(t *testing.T) {
-	tests := []struct {
-		input  string
-		expect CodePipelineEvent
-	}{
-		{
-			input: "testdata/codepipeline-action-execution-stage-change-event.json",
-			expect: CodePipelineEvent{
-				Version:    "0",
-				ID:         "CWE-event-id",
-				DetailType: "CodePipeline Action Execution State Change",
-				Source:     "aws.codepipeline",
-				AccountID:  "123456789012",
-				Time:       time.Date(2017, 04, 22, 3, 31, 47, 0, time.UTC),
-				Region:     "us-east-1",
-				Resources: []string{
-					"arn:aws:codepipeline:us-east-1:123456789012:pipeline:myPipeline",
-				},
-				Detail: CodePipelineEventDetail{
-					Pipeline:    "myPipeline",
-					Version:     1,
-					ExecutionId: "01234567-0123-0123-0123-012345678901",
-					Stage:       "Prod",
-					Action:      "myAction",
-					State:       "STARTED",
-					Region:      "us-west-2",
-					Type: CodePipelineEventDetailType{
-						Owner:    "AWS",
-						Category: "Deploy",
-						Provider: "CodeDeploy",
-						Version:  "1",
-					},
-				},
-			},
-		},
-		{
-			input: "testdata/codepipeline-execution-stage-change-event.json",
-			expect: CodePipelineEvent{
-				Version:    "0",
-				ID:         "CWE-event-id",
-				DetailType: "CodePipeline Stage Execution State Change",
-				Source:     "aws.codepipeline",
-				AccountID:  "123456789012",
-				Time:       time.Date(2017, 04, 22, 3, 31, 47, 0, time.UTC),
-				Region:     "us-east-1",
-				Resources: []string{
-					"arn:aws:codepipeline:us-east-1:123456789012:pipeline:myPipeline",
-				},
-				Detail: CodePipelineEventDetail{
-					Pipeline:    "myPipeline",
-					Version:     1,
-					ExecutionId: "01234567-0123-0123-0123-012345678901",
-					State:       "STARTED",
-				},
-			},
-		},
-		{
-			input: "testdata/codepipeline-execution-state-change-event.json",
-			expect: CodePipelineEvent{
-				Version:    "0",
-				ID:         "CWE-event-id",
-				DetailType: "CodePipeline Pipeline Execution State Change",
-				Source:     "aws.codepipeline",
-				AccountID:  "123456789012",
-				Time:       time.Date(2017, 04, 22, 3, 31, 47, 0, time.UTC),
-				Region:     "us-east-1",
-				Resources: []string{
-					"arn:aws:codepipeline:us-east-1:123456789012:pipeline:myPipeline",
-				},
-				Detail: CodePipelineEventDetail{
-					Pipeline:    "myPipeline",
-					Version:     1,
-					ExecutionId: "01234567-0123-0123-0123-012345678901",
-					State:       "STARTED",
-				},
-			},
-		},
+func TestCodePipeLineJobEventMarshaling(t *testing.T) {
+
+	// read json from file
+	inputJSON, err := ioutil.ReadFile("./testdata/codepipeline-job-event.json")
+	if err != nil {
+		t.Errorf("could not open test file. details: %v", err)
 	}
 
-	for _, testcase := range tests {
-		data, err := ioutil.ReadFile(testcase.input)
-		require.NoError(t, err)
-
-		var actual CodePipelineEvent
-		require.NoError(t, json.Unmarshal(data, &actual))
-
-		require.Equal(t, testcase.expect, actual)
+	// de-serialize into CognitoEvent
+	var inputEvent CodePipelineEvent
+	if err := json.Unmarshal(inputJSON, &inputEvent); err != nil {
+		t.Errorf("could not unmarshal event. details: %v", err)
 	}
+
+	// serialize to json
+	outputJSON, err := json.Marshal(inputEvent)
+	if err != nil {
+		t.Errorf("could not marshal event. details: %v", err)
+	}
+
+	assert.JSONEq(t, string(inputJSON), string(outputJSON))
+}
+
+func TestCodePipelineJobEventMarshalingMalformedJson(t *testing.T) {
+	test.TestMalformedJson(t, CodePipelineEvent{})
 }
